requirements.txt
fastapi
uvicorn[standard]
python-dotenv
twilio
websockets
numpy
